// src/chains/soonsocial/types/v1500.ts
export type Content = Content_None | Content_Other | Content_IPFS

export interface Content_None {
    __kind: 'None'
}

export interface Content_Other {
    __kind: 'Other'
    value: Uint8Array
}

export interface Content_IPFS {
    __kind: 'IPFS'
    value: Uint8Array
}

export type InnerValue = InnerValue_Account | InnerValue_Space | InnerValue_Post

export interface InnerValue_Account {
    __kind: 'Account'
    value: Uint8Array
}

export interface InnerValue_Space {
    __kind: 'Space'
    value: bigint
}

export interface InnerValue_Post {
    __kind: 'Post'
    value: bigint
}
// ...
export interface WhoAndWhen {
    account: Uint8Array
    block: number
    time: bigint
}
// ...
export interface DomainMeta {
    created: WhoAndWhen
    updated: (WhoAndWhen | undefined)
    expiresAt: number
    owner: Uint8Array
    content: Content
    innerValue: (InnerValue | undefined)
    outerValue: (Uint8Array | undefined)
    domainDeposit: bigint
    outerValueDeposit: bigint
}
// ...
export interface Post {
    id: bigint
    created: WhoAndWhen
    edited: boolean
    owner: Uint8Array
    extension: PostExtension
    spaceId: (bigint | undefined)
    content: Content
    hidden: boolean
    upvotesCount: number
    downvotesCount: number
}

export interface Reaction {
    id: bigint
    created: WhoAndWhen
    kind: ReactionKind
}

export interface Role {
    created: WhoAndWhen
    id: bigint
    spaceId: bigint
    disabled: boolean
    expiresAt: (number | undefined)
    content: Content
    permissions: SpacePermission[]
}

export interface Space {
    id: bigint
    created: WhoAndWhen
    edited: boolean
    owner: Uint8Array
    content: Content
    hidden: boolean
    permissions: (SpacePermissions | undefined)
}